requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pinecone==7.3.0
openai==1.3.7
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
requests>=2.32.3
boto3>=1.34.0
jmespath==1.0.1

# Document processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
SpeechRecognition==3.10.0
pydub==0.25.1
